CLI updates and Zigbee send scripts
flightControls-created a crude message to send through the XBEE to the drones

swarmCLI-rearranging. Also telling user which drones are in the network

networkDiscovery-simplified

openBaseStationXBEE-simple function to open a serial connection to the usb xbee

sendMessage-added a direct message and a broadcast message option

zigbeeTest-ignore this file. Simply for debugging

.gitignore-this file will ignore some annoying binary files

README-crude update for steps to start using the scripts